Added svg check status code
