+++ .github/workflows/post_check_svgs_comment.yml
@@ -5,16 +5,17 @@ on:
     types: 
       - completed
 jobs: 
-  post_screenshots_in_comment:
-    name: Post the screenshot
+  post_result_of_svg_check:
+    name: Post the result of the Check SVG Action
     runs-on: ubuntu-18.04
     steps:
       - name: Check if the trigger run worked. If it failed, fail the current run.
-        if: success() && github.event.workflow_run.conclusion != 'success' 
+        if: github.event.workflow_run.conclusion != 'success' 
         uses: cutenode/action-always-fail@v1.0.1
 
       - name: Download workflow artifact
         uses: dawidd6/action-download-artifact@v2.11.0
+        if: success()
         with:
           github_token: ${{ secrets.GITHUB_TOKEN }}
           workflow: peek_icons.yml
@@ -36,13 +37,11 @@ jobs:
 
       - name: Comment on the PR about the result - Success
         uses: jungwinter/comment@v1 # let us comment on a specific PR
-        if: success() && steps.err_message_reader.outputs.content == 'None'
+        if: success() && steps.err_message_reader.outputs.content == '1'
         env:
           MESSAGE: |
             Hi!
-            I'm Devicons' SVG-Checker Bot and I just checked all the SVGs in this branch. 
-
-            Everything looks great. Good job!
+            I'm Devicons' SVG-Checker Bot and everything looks great. Good job!
 
             Have a nice day, 
             SVG-Checker Bot :grin:
@@ -54,7 +53,7 @@ jobs:
 
       - name: Comment on the PR about the result - SVG Error
         uses: jungwinter/comment@v1 # let us comment on a specific PR
-        if: success() && steps.err_message_reader.outputs.content != 'None'
+        if: success() && (steps.err_message_reader.outputs.content != '0' || steps.err_message_reader.outputs.content != '1')
         env:
           MESSAGE: |
             Hi!
@@ -82,7 +81,7 @@ jobs:
 
       - name: Comment on the PR about the result - Failure
         uses: jungwinter/comment@v1 # let us comment on a specific PR
-        if: failure() || cancelled()
+        if: failure() 
         env:
           MESSAGE: |
             Hi!
